move node.rs to new cluster mod
