+++ scylla/src/cluster/mod.rs
@@ -1,6 +1,9 @@
 #[allow(clippy::module_inception)]
 mod cluster;
 pub mod metadata;
+pub(crate) mod node;
 
 pub use cluster::ClusterData;
 pub(crate) use cluster::{use_keyspace_result, Cluster, ClusterNeatDebug};
+
+pub use node::{KnownNode, Node, NodeAddr, NodeRef};
