+++ scylla/src/connection/connection.rs
@@ -51,11 +51,11 @@ use std::{
 #[cfg(feature = "cloud")]
 use crate::cloud::CloudConfig;
 use crate::cluster::metadata::{PeerEndpoint, UntranslatedEndpoint, UntranslatedPeer};
+use crate::cluster::NodeAddr;
 use crate::transport::errors::{ProtocolError, SchemaVersionFetchError, UseKeyspaceProtocolError};
 use crate::transport::iterator::QueryPager;
 use crate::transport::locator::tablets::{RawTablet, TabletParsingError};
 use crate::transport::query_result::QueryResult;
-use crate::transport::NodeAddr;
 
 use crate::batch::{Batch, BatchStatement};
 use crate::connection::Compression;
@@ -2481,9 +2481,9 @@ mod tests {
 
     use super::{open_connection, ConnectionConfig};
     use crate::cluster::metadata::UntranslatedEndpoint;
+    use crate::cluster::node::ResolvedContactPoint;
     use crate::query::Query;
     use crate::test_utils::setup_tracing;
-    use crate::transport::node::ResolvedContactPoint;
     use crate::utils::test_utils::{unique_keyspace_name, PerformDDL};
     use crate::SessionBuilder;
     use futures::{StreamExt, TryStreamExt};
