+++ scylla/src/session/session.rs
@@ -36,6 +36,11 @@ use tokio::time::timeout;
 use tracing::{debug, error, trace, trace_span, Instrument};
 use uuid::Uuid;
 
+#[cfg(feature = "cloud")]
+use crate::cluster::node::CloudEndpoint;
+use crate::cluster::node::Node;
+use crate::cluster::node::{InternalKnownNode, KnownNode};
+use crate::cluster::NodeRef;
 use crate::cluster::{Cluster, ClusterData, ClusterNeatDebug};
 #[cfg(feature = "ssl")]
 use crate::connection::SslConfig;
@@ -59,17 +64,12 @@ use crate::transport::iterator::QueryPager;
 use crate::transport::iterator::{LegacyRowIterator, PreparedIteratorConfig};
 use crate::transport::load_balancing::{self, RoutingInfo};
 use crate::transport::metrics::Metrics;
-#[cfg(feature = "cloud")]
-use crate::transport::node::CloudEndpoint;
-use crate::transport::node::Node;
-use crate::transport::node::{InternalKnownNode, KnownNode};
 use crate::transport::partitioner::PartitionerName;
 use crate::transport::query_result::MaybeFirstRowError;
 use crate::transport::query_result::QueryResult;
 use crate::transport::query_result::RowsError;
 use crate::transport::retry_policy::{QueryInfo, RetryDecision, RetrySession};
 use crate::transport::speculative_execution;
-use crate::transport::NodeRef;
 use crate::{
     batch::{Batch, BatchStatement},
     statement::StatementConfig,
@@ -957,7 +957,7 @@ where
     /// # use std::error::Error;
     /// # async fn check_only_compiles() -> Result<(), Box<dyn Error>> {
     /// use scylla::{Session, SessionConfig};
-    /// use scylla::transport::KnownNode;
+    /// use scylla::cluster::KnownNode;
     ///
     /// let mut config = SessionConfig::new();
     /// config.known_nodes.push(KnownNode::Hostname("127.0.0.1:9042".to_string()));
