+++ scylla/src/transport/iterator.rs
@@ -31,6 +31,7 @@ use crate::deserialize::DeserializeOwnedRow;
 use crate::session::RequestSpan;
 
 use crate::cluster::ClusterData;
+use crate::cluster::NodeRef;
 use crate::connection::{Connection, NonErrorQueryResponse, QueryResponse};
 use crate::frame::response::{
     result,
@@ -42,7 +43,6 @@ use crate::transport::errors::ProtocolError;
 use crate::transport::load_balancing::{self, RoutingInfo};
 use crate::transport::metrics::Metrics;
 use crate::transport::retry_policy::{QueryInfo, RetryDecision, RetrySession};
-use crate::transport::NodeRef;
 use tracing::{trace, trace_span, warn, Instrument};
 use uuid::Uuid;
