+++ scylla/src/transport/locator/test.rs
@@ -6,11 +6,11 @@ use uuid::Uuid;
 use super::tablets::TabletsInfo;
 use super::{ReplicaLocator, ReplicaSet};
 use crate::cluster::metadata::{Keyspace, Metadata, Peer, Strategy};
+use crate::cluster::Node;
+use crate::cluster::{NodeAddr, NodeRef};
 use crate::connection::PoolConfig;
 use crate::routing::Token;
 use crate::test_utils::setup_tracing;
-use crate::transport::Node;
-use crate::transport::{NodeAddr, NodeRef};
 
 use std::collections::HashSet;
 use std::sync::Arc;
