+++ scylla/src/transport/mod.rs
@@ -7,7 +7,6 @@ pub mod legacy_query_result;
 pub mod load_balancing;
 pub mod locator;
 pub(crate) mod metrics;
-pub(crate) mod node;
 pub mod partitioner;
 pub mod query_result;
 pub mod retry_policy;
@@ -16,5 +15,3 @@ pub mod speculative_execution;
 pub use crate::frame::Authenticator;
 pub use execution_profile::ExecutionProfile;
 pub use scylla_cql::frame::request::query::{PagingState, PagingStateResponse};
-
-pub use node::{KnownNode, Node, NodeAddr, NodeRef};
